requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
scikit-learn==1.3.2
tensorflow==2.15.0
joblib==1.3.2
scapy==2.5.0
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2